debug

+++ runFO1DSimulation.py
@@ -68,7 +68,7 @@ a = np.sum(np.abs(q) <= q_max)
 
 
 if len(q)>a:
-    q = q[np.ceil(n_sample/2+1-(a-1)/2):np.floor(n_sample/2+1+(a+1)/2)]
+    q = q[int(np.ceil(n_sample/2+1-(a-1)/2)):int(np.floor(n_sample/2+1+(a+1)/2))]
     F_wave_obj = F_wave_obj[np.ceil(n_sample / 2 + 1 - (a - 1) / 2):np.floor(n_sample / 2 + 1 + (a + 1) / 2)]
 
 Q, QQ = np.meshgrid(q, q)
